+++ src/ocfl/consts.rs
@@ -28,16 +28,18 @@ pub const HASHED_NTUPLE_OBJECT_ID_LAYOUT_EXTENSION: &str =
 pub const HASHED_NTUPLE_LAYOUT_EXTENSION: &str = "0004-hashed-n-tuple-storage-layout";
 pub const MUTABLE_HEAD_EXTENSION: &str = "0005-mutable-head";
 pub const FLAT_OMIT_PREFIX_LAYOUT_EXTENSION: &str = "0006-flat-omit-prefix-storage-layout";
+pub const NTUPLE_OMIT_PREFIX_LAYOUT_EXTENSION: &str = "0007-n-tuple-omit-prefix-storage-layout";
 pub const ROCFL_STAGING_EXTENSION: &str = "rocfl-staging";
 pub const ROCFL_LOCKS_EXTENSION: &str = "rocfl-locks";
 
 pub static SUPPORTED_EXTENSIONS: Lazy<HashSet<&str>> = Lazy::new(|| {
-    let mut set = HashSet::with_capacity(6);
+    let mut set = HashSet::with_capacity(8);
     set.insert(FLAT_DIRECT_LAYOUT_EXTENSION);
     set.insert(HASHED_NTUPLE_OBJECT_ID_LAYOUT_EXTENSION);
     set.insert(HASHED_NTUPLE_LAYOUT_EXTENSION);
     set.insert(MUTABLE_HEAD_EXTENSION);
     set.insert(FLAT_OMIT_PREFIX_LAYOUT_EXTENSION);
+    set.insert(NTUPLE_OMIT_PREFIX_LAYOUT_EXTENSION);
     set.insert(ROCFL_STAGING_EXTENSION);
     set.insert(ROCFL_LOCKS_EXTENSION);
     set