+++ src/ocfl/store/fs.rs
@@ -1160,6 +1160,7 @@ fn write_layout_config(root: impl AsRef<Path>, layout: &StorageLayout) -> Result
         LayoutExtensionName::HashedNTupleObjectIdLayout => specs::EXT_0003_SPEC,
         LayoutExtensionName::HashedNTupleLayout => specs::EXT_0004_SPEC,
         LayoutExtensionName::FlatOmitPrefixLayout => specs::EXT_0006_SPEC,
+        LayoutExtensionName::NTupleOmitPrefixLayout => specs::EXT_0007_SPEC,
     };
 
     write!(