+++ CHANGELOG.md
@@ -10,6 +10,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Support for layout extension [0007-n-tuple-omit-prefix-storage-layout](https://ocfl.github.io/extensions/0007-n-tuple-omit-prefix-storage-layout.html)
+
 ### Changed
 
 - Upgraded to [digest](https://crates.io/crates/digest) `0.10.0`, which